+++ src/Extension/Symfony/Config/ChainableFileLocatorInterface.php
@@ -1,0 +1,23 @@
+<?php
+
+/*
+ * This file is part of the Symfony Puli bridge.
+ *
+ * (c) Bernhard Schussek <bschussek@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Webmozart\Puli\Extension\Symfony\Config;
+
+use Symfony\Component\Config\FileLocatorInterface;
+
+/**
+ * @since  1.0
+ * @author Bernhard Schussek <bschussek@gmail.com>
+ */
+interface ChainableFileLocatorInterface extends FileLocatorInterface
+{
+    public function supports($path);
+}